Reduce size of `AttachmentDesc` by moving stuff around and changing underlying type of some enums

+++ src/rendering/TemporalBlending.hpp
@@ -16,7 +16,7 @@ class GBuffer;
 class Texture;
 struct RenderSetup;
 
-enum TextureFormat : uint32;
+enum TextureFormat : uint8;
 
 static constexpr double DefaultTemporalBlendingFeedback = 0.8;
